# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tutor/practice:
    post:
      operationId: generatePractice
      tags: [tutor]
      summary: Generate practice questions from tutor response context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePracticeBody"
      responses:
        "200":
          description: Practice questions output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
  /tutor/diagnostic:
    post:
      operationId: runDiagnostic
      tags: [tutor]
      summary: Run trace-back diagnostic evaluation on a student's answer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunDiagnosticBody"
      responses:
        "200":
          description: Diagnostic evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
  /tutor/invoke:
    post:
      operationId: invokeTutor
      tags: [tutor]
      summary: Invoke the adversarial AI tutor workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvokeTutorBody"
      responses:
        "200":
          description: Tutor response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
...
components:
  schemas:
...
    GeminiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    InvokeTutorBody:
      type: object
      properties:
        userInput:
          type: string
        targetExam:
          type: string
        subject:
          type: string
        pedagogyStyle:
          type: string
          enum: [hinglish, english, mnemonic, simplified, standard, abstract]
        studentAnswer:
          type: string
        questionMode:
          type: boolean
        conversationId:
          type: integer
      required:
        - userInput
        - targetExam
    TutorResponse:
      type: object
      properties:
        response:
          type: string
        cacheHit:
          type: boolean
        gradingPassed:
          type: boolean
        mnemonics:
          type: array
          items:
            type: string
        weakTopics:
          type: array
          items:
            type: string
        revisionSuggestions:
          type: array
          items:
            type: string
        conversationId:
          type: integer
        messageId:
          type: integer
      required:
        - response
        - cacheHit
        - gradingPassed
...
    GeneratePracticeBody:
      type: object
      properties:
        context:
          type: string
        targetExam:
          type: string
        language:
          type: string
        model:
          type: string
      required:
        - context
        - targetExam
    PracticeResponse:
      type: object
      properties:
        questions:
          type: string
      required:
        - questions
    RunDiagnosticBody:
      type: object
      properties:
        studentAnswer:
          type: string
        reference:
          type: string
        targetExam:
          type: string
        language:
          type: string
        model:
          type: string
      required:
        - studentAnswer
        - reference
        - targetExam
    DiagnosticResponse:
      type: object
      properties:
        evaluation:
          type: string
        score:
          type: number
      required:
        - evaluation
        - score